fix issue with overrides by renaming the chart
