+++ module-chart/charts/backend/Chart.yaml
@@ -1,4 +1,4 @@
 apiVersion: v2
-name: nats
+name: backend
 description: Helm chart for nats
 version: 0.17.3
