+++ module-chart/charts/backend/templates/statefulset.yaml
@@ -1,4 +1,4 @@
-{{if eq .Values.backend.type "nats"}}
+{{if eq .Values.type "nats"}}
 ---
 apiVersion: apps/v1
 kind: StatefulSet
