+++ module-chart/charts/backend/values.yaml
@@ -266,5 +266,4 @@ exporter:
     # interval:
     # scrapeTimeout:
 
-backend:
-  type: "nats"
+type: "nats"
